+++ packages/spacecat-shared-data-access/src/service/index.js
@@ -12,10 +12,12 @@
 
 import { createClient } from '@adobe/spacecat-shared-dynamo';
 import { auditFunctions } from './audits/index.js';
+import { keyEventFunctions } from './key-events/index.js';
 import { siteFunctions } from './sites/index.js';
 import { siteCandidateFunctions } from './site-candidates/index.js';
 import { organizationFunctions } from './organizations/index.js';
 import { configurationFunctions } from './configurations/index.js';
+import { siteTopPagesFunctions } from './site-top-pages/index.js';
 
 /**
  * Creates a data access object.
@@ -23,8 +25,8 @@ import { configurationFunctions } from './configurations/index.js';
  * @param {{pkAllSites: string, pkAllLatestAudits: string, indexNameAllLatestAuditScores: string,
  * tableNameAudits: string,tableNameLatestAudits: string, indexNameAllSitesOrganizations: string,
  * tableNameSites: string, tableNameOrganizations: string, indexNameAllSites: string,
- * indexNameAllOrganizations: string, indexNameAllOrganizationsByImsOrgId: string,
- * pkAllOrganizations: string}} config configuration
+ * tableNameSiteTopPages: string, indexNameAllOrganizations: string,
+ * indexNameAllOrganizationsByImsOrgId: string, pkAllOrganizations: string}} config configuration
  * @param {Logger} log logger
  * @returns {object} data access object
  */
@@ -32,16 +34,20 @@ export const createDataAccess = (config, log = console) => {
   const dynamoClient = createClient(log);
 
   const auditFuncs = auditFunctions(dynamoClient, config, log);
+  const keyEventFuncs = keyEventFunctions(dynamoClient, config, log);
   const siteFuncs = siteFunctions(dynamoClient, config, log);
   const siteCandidateFuncs = siteCandidateFunctions(dynamoClient, config, log);
   const organizationFuncs = organizationFunctions(dynamoClient, config, log);
   const configurationFuncs = configurationFunctions(dynamoClient, config);
+  const siteTopPagesFuncs = siteTopPagesFunctions(dynamoClient, config);
 
   return {
     ...auditFuncs,
+    ...keyEventFuncs,
     ...siteFuncs,
     ...siteCandidateFuncs,
     ...organizationFuncs,
     ...configurationFuncs,
+    ...siteTopPagesFuncs,
   };
 };