+++ packages/spacecat-shared-data-access/test/unit/service/index.test.js
@@ -66,6 +66,18 @@ describe('Data Access Object Tests', () => {
     'updateConfiguration',
   ];
 
+  const siteTopPagesFunctions = [
+    'getTopPagesForSite',
+    'addSiteTopPage',
+    'removeSiteTopPages',
+  ];
+
+  const keyEventsFunctions = [
+    'createKeyEvent',
+    'getKeyEventsForSite',
+    'removeKeyEvent',
+  ];
+
   let dao;
 
   before(() => {
@@ -90,13 +102,27 @@ describe('Data Access Object Tests', () => {
     });
   });
 
+  it('contains all known site top pages functions', () => {
+    siteTopPagesFunctions.forEach((funcName) => {
+      expect(dao).to.have.property(funcName);
+    });
+  });
+
+  it('contains all known key events functions', () => {
+    keyEventsFunctions.forEach((funcName) => {
+      expect(dao).to.have.property(funcName);
+    });
+  });
+
   it('does not contain any unexpected functions', () => {
     const expectedFunctions = new Set([
       ...auditFunctions,
+      ...keyEventsFunctions,
       ...siteFunctions,
       ...siteCandidateFunctions,
       ...organizationFunctions,
-      ...configurationFunctions]);
+      ...configurationFunctions,
+      ...siteTopPagesFunctions]);
     Object.keys(dao).forEach((funcName) => {
       expect(expectedFunctions).to.include(funcName);
     });