+++ packages/spacecat-shared-rum-api-client/src/index.js
@@ -21,6 +21,7 @@ const APIS = {
   NOT_FOUND_DASHBOARD_UI: 'https://data.aem.live/404-reports',
   RUM_DASHBOARD: 'https://helix-pages.anywhere.run/helix-services/run-query@v3/rum-dashboard',
   DOMAIN_LIST: 'https://helix-pages.anywhere.run/helix-services/run-query@v3/dash/domain-list',
+  RUM_404: 'https://helix-pages.anywhere.run/helix-services/run-query@v3/rum-404',
   RUM_SOURCES: 'https://helix-pages.anywhere.run/helix-services/run-query@v3/rum-sources',
   RUM_EXPERIMENTS: 'https://helix-pages.anywhere.run/helix-services/run-query@v3/rum-experiments',
 };
@@ -37,15 +38,16 @@ export const RUM_DEFAULT_PARAMS = {
   limit: 101,
 };
 
-export const NOT_FOUND_DEFAULT_PARAMS = {
+export const CONVERSION_DEFAULT_PARAMS = {
   ...RUM_DEFAULT_PARAMS,
-  checkpoint: 404,
+  checkpoint: 'convert',
+  aggregate: false,
 };
 
 export const create404URL = (params = {}) => createUrl(
-  APIS.RUM_SOURCES,
+  APIS.RUM_404,
   {
-    ...NOT_FOUND_DEFAULT_PARAMS, ...params,
+    ...RUM_DEFAULT_PARAMS, ...params,
   },
 );
 
@@ -60,6 +62,13 @@ export const createExperimentationURL = (params = {}) => createUrl(
   APIS.RUM_EXPERIMENTS,
   {
     ...RUM_DEFAULT_PARAMS, ...params,
+  },
+);
+
+export const createConversionURL = (params = {}) => createUrl(
+  APIS.RUM_SOURCES,
+  {
+    ...CONVERSION_DEFAULT_PARAMS, ...params,
   },
 );
 
@@ -150,6 +159,10 @@ export default class RUMAPIClient {
 
   async getExperimentationData(params = {}) {
     return sendRequest(createExperimentationURL({ ...params, domainkey: this.domainkey }));
+  }
+
+  async getConversionData(params = {}) {
+    return sendRequest(createConversionURL({ ...params, domainkey: this.domainkey }));
   }
 
   async get404Sources(params = {}) {