+++ packages/spacecat-shared-rum-api-client/test/create-urls.test.js
@@ -14,7 +14,9 @@
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import nock from 'nock';
-import { create404URL, createRUMURL, createExperimentationURL } from '../src/index.js';
+import {
+  create404URL, createRUMURL, createExperimentationURL, createConversionURL,
+} from '../src/index.js';
 
 chai.use(chaiAsPromised);
 const { expect } = chai;
@@ -26,7 +28,7 @@ describe('create urls tests', () => {
 
   it('returns the URL to call the get404Sources', () => {
     expect(create404URL({ url: 'http://spacecar.com' }))
-      .to.eql('https://helix-pages.anywhere.run/helix-services/run-query@v3/rum-sources?interval=7&offset=0&limit=101&checkpoint=404&url=http%3A%2F%2Fspacecar.com');
+      .to.eql('https://helix-pages.anywhere.run/helix-services/run-query@v3/rum-404?interval=7&offset=0&limit=101&url=http%3A%2F%2Fspacecar.com');
   });
 
   it('returns the URL to call the getRUMDashboard', () => {
@@ -37,5 +39,10 @@ describe('create urls tests', () => {
   it('returns the URL to get the Experimentation data', () => {
     expect(createExperimentationURL({ url: 'http://spacecat.com' }))
       .to.eql('https://helix-pages.anywhere.run/helix-services/run-query@v3/rum-experiments?interval=7&offset=0&limit=101&url=http%3A%2F%2Fspacecat.com');
+  });
+
+  it('returns the URL to get the Conversion data', () => {
+    expect(createConversionURL({ url: 'http://spacecat.com' }))
+      .to.eql('https://helix-pages.anywhere.run/helix-services/run-query@v3/rum-sources?interval=7&offset=0&limit=101&checkpoint=convert&aggregate=false&url=http%3A%2F%2Fspacecat.com');
   });
 });